Summary: portability fixes
Use inet_pton if getnameinfo isn't available. (Cygwin)
Don't check for socklen_t, we always screw the check up, and all our target platforms have it.  (Cygwin/MacOS X 10.3/Solaris)
Use fcntl with O_NONBLOCK instead of the MSG_DONTWAIT flag.  (Cygwin)
Rename parameter s_addr to s_netaddr in mud/telnet.h and mud/telnet.cc, since s_addr can be a freaky macro.  (Solaris)
Make md5.cc use uint32_t from common/types.h.  (Solaris)
Move md5.cc from src/mud to src/common.
Fixup common/types.h to use standard int32_t and friends if stdint.h or inttypes.h is available.
NOTE: I have compiled and connected to AweMUD successfully on a sparc Solaris machine... only to have it crash almost immediately.  Not sure if the bug is in libgc or AweMUD itself yet.

+++ include/scriptix/value.h
@@ -30,7 +30,6 @@
 
 #include "common/gcbase.h"
 #include "common/awestr.h"
-#include "scriptix/sysdep.h"
 
 namespace Scriptix {
 
@@ -62,7 +61,7 @@ class Value : public GCType::GC {
 	Value (long value);
 
 	inline Value (const char* s_value) { *this = Value(String(s_value)); }
-	inline Value (intptr_t s_value) { *this = Value((long)s_value); }
+	inline Value (int s_value) { *this = Value((long)s_value); }
 
 	// get IValue*
 	inline operator IValue* () { return value; }
